+++ src/features/member/components/new/BirthInputForm.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useController, useFormContext } from 'react-hook-form';
 import { useRouter } from 'next/navigation';
 
@@ -8,42 +8,25 @@ import BirthIcon from '@/assets/icons/birth-icon.svg';
 import { Button } from '@/components';
 import { MAX_DATE_LENGTH_UNTIL_DAY } from '@/constants';
 import { DateInput } from '@/features/goal/components/new/DateInput';
-import { useToast } from '@/hooks/useToast';
+import { useValidBirth } from '@/hooks/useValidBirth';
 
 import type { NewMemberFormValues } from '../../types';
 
 import FormLayout from './FormLayout';
 
 export const BirthInputForm = () => {
-  const toast = useToast();
   const router = useRouter();
   const { register, getValues, control } = useFormContext<NewMemberFormValues>();
   const { field } = useController({ name: 'birth', control });
   const { onChange, value } = field;
+  const isValidBirth = useValidBirth(value);
   const { nickname } = getValues();
-  const [isValidBirth, setIsValidBirth] = useState(false);
 
   useEffect(() => {
     if (!nickname) {
       router.push('/member/new/nickname');
     }
   }, []);
-
-  useEffect(() => {
-    if (value) {
-      if (value.length === MAX_DATE_LENGTH_UNTIL_DAY) {
-        const today = new Date();
-        const inputValue = new Date(value);
-
-        if (today < inputValue) {
-          toast.warning('올바른 생년월일을 입력해주세요.');
-          setIsValidBirth(false);
-          return;
-        }
-      }
-      setIsValidBirth(value.length === MAX_DATE_LENGTH_UNTIL_DAY);
-    }
-  }, [value]);
 
   return (
     <FormLayout
